Legg til steg for å vurdere arbeidsland med sokkel/skip
