+++ src/felleskomponenter/sokkelskipliste/index.js
@@ -1,0 +1,3 @@
+import Sokkelskipliste from './sokkelskipliste';
+
+export default Sokkelskipliste;
