+++ src/felleskomponenter/sokkelskipliste/sokkelskipenkelt.less
@@ -1,0 +1,8 @@
+.sokkelSkip__liste__rad {
+  border-bottom: 1px solid #cccccc;
+  margin: 1em 0;
+
+  &:last-of-type {
+    border-bottom: 0;
+  }
+}
