+++ src/felleskomponenter/stegvelger/stegKomponenter/vurderingSokkelSkip.less
@@ -4,13 +4,4 @@
   .sokkelSkip__varsel {
     margin: 1em 0;
   }
-
-  .sokkelSkip__liste__rad {
-    border-bottom: 1px solid #cccccc;
-    margin: 1em 0;
-
-    &:last-of-type {
-      border-bottom: 0;
-    }
-  }
 }
