+++ src/felleskomponenter/stegvelger/stegKomponenter/vurderingVurderarbeidsland.less
@@ -1,0 +1,5 @@
+@import '../../../constants';
+
+.borderBottom {
+  border-bottom: 1px solid @gray;
+}
