+++ src/felleskomponenter/stegvelger/stegMotor/typer.js
@@ -46,6 +46,7 @@ export const STEG = {
   VEDTAK: 'VEDTAK',
   ENDRET_PERIODE: 'ENDRE_PERIODE',
   VIDERESEND: 'VIDERESEND',
+  VURDER_ARBEIDSLAND: 'VURDER_ARBEIDSLAND',
   VURDER_UTPEKING: 'VURDER_UTPEKING',
   GODKJENN_UTPEKING: 'GODKJENN_UTPEKING',
   AVSLAA_UTPEKING: 'AVSLAA_UTPEKING',
